mas_apps/search/requirements.txt
openai>=1.40.0
PyYAML>=6.0
requests>=2.31.0
pandas>=2.0.0
pyarrow>=14.0.0
tqdm>=4.66.0
huggingface_hub>=0.23.0
fastapi>=0.111.0
uvicorn>=0.30.0
opensearch-py>=2.6.0
